doc:
Former-commit-id: d44fc4a7513bde665bc5431d08217a1b54a2b364 [formerly 3d3a18837a9f8cc6784380be6c6e3389249afd30]
Former-commit-id: dcf734d359ef27b4856cb2cd92e411a76bb627a6

+++ SRACore/util/operator.py
@@ -517,7 +517,7 @@ class Operator:
                 return Box(left, top, width, height)
         return None
 
-    def wait_ocr(self, text: str, timeout: int = 10, interval: float = 0.2, confidence=0.9, *args, **kwargs) -> Box | None:
+    def wait_ocr(self, text: str, timeout: float = 10, interval: float = 0.2, confidence=0.9, *args, **kwargs) -> Box | None:
         """
         等待OCR识别到指定文本
         :param text: 要识别的文本
